requirements.txt
requests==2.32.3
pandas==2.2.2
google-cloud-storage==2.18.2
python-dotenv==1.0.1
pyarrow==17.0.0
pytest==8.3.2
requests-mock==1.12.1